Modify BlogBlock, Footer, create post button, pagination, aboutus page, blogs page, category page and navbar
