+++ blog-app/src/components/Footer.css
@@ -1,34 +1,22 @@
 .backgound{
-    background-color: black;
+    background-color: rgba(0, 0, 0, 0.8);
 }
 
 .footer-container {
     display: flex;
-    justify-content: space-between;
-    align-items: center;
     
 }
 
 .logo {
-    display: flex;
-    justify-content: flex-start;
+
 }
 
 .group {
     color: white;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    margin-right: 600px;
-    margin-bottom: 5%;
+    
 }
 
 .usth {
-    width: 400px; /* Điều chỉnh kích thước logo USTH */
+    width: 30%;
     height: auto;
-    margin: 20px;
-}
-
-.team{
-    color: white;
 }
