+++ blog-app/src/components/Footer.js
@@ -10,7 +10,6 @@ function Footer() {
                     <img src={usth} alt='USTH' className='usth' />
                 </div>
                 <div className='group'>
-                    <div className='team'>
                         <h1>GROUP 10</h1>
                         <p>-------------------------------------------------</p>
                         <p>Pham Duc Khiem:          BA12-095</p>
@@ -19,7 +18,7 @@ function Footer() {
                         <p>Tran Ngoc Viet Anh:      BA12-003</p>
                         <p>Nguyen Dinh Hai:         BA12-068</p>
                         <p>Ngo Huyen Anh:           BA12-006</p>
-                    </div>
+                    
                     
                 </div>
             </div>
