+++ blog-app/src/components/Iconplus.css
@@ -1,19 +1,19 @@
-.icon {
+.create_post_button {
     display: flex;
     justify-content: center;
     align-items: center;
     position: fixed;
     bottom: 20px;
-    right: 20px;
-    width: 100px;
-    height: 100px;
+    right: 50px;
+    width: 80px;
+    height: 80px;
     border-radius: 40%;
-    background-color:paleturquoise;
+    background-color: rgba(0, 0, 0, 0.8);
+    font-size: 30px;
+    color: #fff;
+    box-shadow: 4px 4px 8px rgba(0, 0, 0, 0.65);
 }
 
-.icon .plus {
-    color: white;
-    width: 50%;
-    height: 50%;
-    
+.create_post_button:hover {
+    opacity: 0.8;
 }
