--- blog-app/src/components/Iconplus.js
@@ -1,13 +1,0 @@
-import './Iconplus.css'
-import pen from './assets/pen.png'
-
-function Iconplus() {
-    return (
-        <button className='icon'>
-            
-            <img src={pen} alt='' className='plus'/>
-        </button>
-    );
-}
-
-export default Iconplus;
