+++ blog-app/src/pages/Blogs.css
@@ -1,9 +1,25 @@
-.blog-blocks {
+.blog_container {
     display: flex;
-    gap: 20px;
-    flex-wrap: nowrap;
+    justify-content: center;
+    margin-top: 100px;
 }
 
-.line{
-    margin: 20px;
+.blog_content {
+    width: 80%;
+}
+
+.title_blogs {
+    border-bottom: 1px solid rgba(0, 0, 0, 0.4) ;
+    padding: 20px 0;
+    font-size: 26px;
+}
+
+
+
+.blog-blocks {
+    display: flex;  
+    flex-direction: row-reverse;
+    justify-content: flex-end;
+    flex-wrap: wrap;
+    margin: 30px 0;
 }
