+++ blog-app/src/pages/Categories.jsx
@@ -2,6 +2,12 @@ import './Categories.css';
 import BlogBlock from '../components/BlogBlock';
 import NavBar from './NavBar';
 import Footer from '../components/Footer';
+import Pagination from '../components/Pagination';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+
+
+
+
 
 function Categories(){
     return(
@@ -11,9 +17,7 @@ function Categories(){
             </header>
             <main className='category_content'>
                 <div className="category_content_container">
-                    <div className='category_title'>
-                        <h2 >Type I</h2>
-                    </div>
+                    <h2 className='category_title' >Type I</h2>
                     <div className='category_blogblock_container'>
                         <div className="blogblock_container">
                             <BlogBlock/>
@@ -22,16 +26,13 @@ function Categories(){
                             <BlogBlock/>
                             <BlogBlock/>
                             <BlogBlock/>
-
+                            <BlogBlock/>
+                            <BlogBlock/>
+                            <BlogBlock/>
+                            <BlogBlock/>
                         </div>
 
-                        <div class='pagination'>
-                        <a className='pagination_direction previous' href='#'>&laquo; Prev</a>
-                            <a className='active' href='#'>1</a>
-                            <a href='#'>2</a>
-                            <a className='pagination_direction next' href='#'>Next &raquo;</a>
-
-                        </div>
+                        < Pagination/>
                     </div>
                 </div>
             </main>
